fix(settings): wire up app reset in data settings
The reset button's onReset was a no-op stub, so nothing happened. Port the
full reset (books DB, search indexes, user data, custom DB location) into
DataSettingsViewModel.resetApp() and wire the button to it, then restart.
Remove the now-dead ResetApp event/handler/resetDone from general settings
and update the related tests.
