+++ SeforimApp/src/jvmMain/kotlin/io/github/kdroidfilter/seforimapp/features/settings/data/DataSettingsViewModel.kt
@@ -6,8 +6,12 @@ import dev.zacsweers.metro.ContributesIntoMap
 import dev.zacsweers.metro.Inject
 import dev.zacsweers.metrox.viewmodel.ViewModelKey
 import io.github.kdroidfilter.platformtools.appmanager.restartApplication
+import io.github.kdroidfilter.seforimapp.core.settings.AppSettings
 import io.github.kdroidfilter.seforimapp.framework.database.getUserSettingsDatabasePath
 import io.github.kdroidfilter.seforimapp.framework.di.AppScope
+import io.github.vinceglb.filekit.FileKit
+import io.github.vinceglb.filekit.databasesDir
+import io.github.vinceglb.filekit.path
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -86,6 +90,55 @@ class DataSettingsViewModel : ViewModel() {
             } catch (e: Exception) {
                 _state.update { it.copy(isImporting = false, importFailed = true) }
             }
+        }
+    }
+
+    /**
+     * Wipes everything: books database, search indexes and all personal data (notes, highlights,
+     * session) — both the managed databases directory and any custom DB location — then clears the
+     * settings and restarts the app.
+     */
+    fun resetApp() {
+        viewModelScope.launch(Dispatchers.IO) {
+            val dbDir = File(FileKit.databasesDir.path)
+            // Read the custom DB path before clearing settings (clearAll() drops it).
+            val customDbPath = runCatching { AppSettings.getDatabasePath() }.getOrNull()
+
+            AppSettings.clearAll()
+
+            // Delete every file/directory in the managed databases directory (this also holds the
+            // user settings DB with notes and highlights).
+            if (dbDir.exists()) {
+                dbDir.listFiles()?.forEach { file ->
+                    runCatching {
+                        if (file.isDirectory) file.deleteRecursively() else file.delete()
+                    }
+                }
+            }
+
+            // Also clean a custom DB location, if the user pointed the books DB elsewhere.
+            if (!customDbPath.isNullOrBlank()) {
+                val customDbFile = File(customDbPath)
+                val customBaseDir = customDbFile.parentFile
+                runCatching { if (customDbFile.exists()) customDbFile.delete() }
+                if (customBaseDir != null && customBaseDir.exists() && customBaseDir != dbDir) {
+                    listOf(
+                        customDbFile.name + ".lucene",
+                        customDbFile.name + ".lookup.lucene",
+                        "lexical.db",
+                        "catalog.pb",
+                        "release_info.txt",
+                    ).forEach { name ->
+                        val f = File(customBaseDir, name)
+                        if (f.exists()) {
+                            runCatching { if (f.isDirectory) f.deleteRecursively() else f.delete() }
+                        }
+                    }
+                }
+            }
+
+            _state.update { it.copy(resetDone = true) }
+            restartApplication()
         }
     }
 }
