+++ SeforimApp/src/jvmMain/kotlin/io/github/kdroidfilter/seforimapp/features/settings/general/GeneralSettingsViewModel.kt
@@ -5,17 +5,12 @@ import androidx.lifecycle.viewModelScope
 import dev.zacsweers.metro.ContributesIntoMap
 import dev.zacsweers.metro.Inject
 import dev.zacsweers.metrox.viewmodel.ViewModelKey
-import io.github.kdroidfilter.platformtools.appmanager.restartApplication
 import io.github.kdroidfilter.seforimapp.core.settings.AppSettings
 import io.github.kdroidfilter.seforimapp.framework.di.AppScope
-import io.github.vinceglb.filekit.FileKit
-import io.github.vinceglb.filekit.databasesDir
-import io.github.vinceglb.filekit.path
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.stateIn
-import java.io.File
 
 @ContributesIntoMap(AppScope::class)
 @ViewModelKey
@@ -25,16 +20,14 @@ class GeneralSettingsViewModel : ViewModel() {
     private val closeTree = MutableStateFlow(AppSettings.getCloseBookTreeOnNewBookSelected())
     private val persist = MutableStateFlow(AppSettings.isPersistSessionEnabled())
     private val keepAwake = MutableStateFlow(AppSettings.isKeepScreenAwakeOnBookEnabled())
-    private val resetDone = MutableStateFlow(false)
 
     val state =
-        combine(dbPath, closeTree, persist, keepAwake, resetDone) { path, c, p, k, r ->
+        combine(dbPath, closeTree, persist, keepAwake) { path, c, p, k ->
             GeneralSettingsState(
                 databasePath = path,
                 closeTreeOnNewBook = c,
                 persistSession = p,
                 keepScreenAwakeOnBook = k,
-                resetDone = r,
             )
         }.stateIn(
             viewModelScope,
@@ -44,7 +37,6 @@ class GeneralSettingsViewModel : ViewModel() {
                 closeTreeOnNewBook = closeTree.value,
                 persistSession = persist.value,
                 keepScreenAwakeOnBook = keepAwake.value,
-                resetDone = resetDone.value,
             ),
         )
 
@@ -61,61 +53,6 @@ class GeneralSettingsViewModel : ViewModel() {
             is GeneralSettingsEvents.SetKeepScreenAwakeOnBook -> {
                 AppSettings.setKeepScreenAwakeOnBookEnabled(event.value)
                 keepAwake.value = event.value
-            }
-            is GeneralSettingsEvents.ResetApp -> {
-                // Get the databases directory
-                val dbDir = File(FileKit.databasesDir.path)
-
-                // Get custom DB path BEFORE clearing settings
-                val customDbPath = runCatching { AppSettings.getDatabasePath() }.getOrNull()
-
-                // Clear settings first
-                AppSettings.clearAll()
-
-                // Delete all files and directories in the databases directory
-                if (dbDir.exists()) {
-                    dbDir.listFiles()?.forEach { file ->
-                        runCatching {
-                            if (file.isDirectory) {
-                                file.deleteRecursively()
-                            } else {
-                                file.delete()
-                            }
-                        }
-                    }
-                }
-
-                // Also delete database files in custom path (if user selected a different location)
-                if (!customDbPath.isNullOrBlank()) {
-                    val customDbFile = File(customDbPath)
-                    val customBaseDir = customDbFile.parentFile
-
-                    // Delete the database file
-                    runCatching { if (customDbFile.exists()) customDbFile.delete() }
-
-                    // Delete related files in the custom directory
-                    if (customBaseDir != null && customBaseDir.exists() && customBaseDir != dbDir) {
-                        val relatedPatterns =
-                            listOf(
-                                customDbFile.name + ".lucene",
-                                customDbFile.name + ".lookup.lucene",
-                                "lexical.db",
-                                "catalog.pb",
-                                "release_info.txt",
-                            )
-                        relatedPatterns.forEach { name ->
-                            val f = File(customBaseDir, name)
-                            if (f.exists()) {
-                                runCatching {
-                                    if (f.isDirectory) f.deleteRecursively() else f.delete()
-                                }
-                            }
-                        }
-                    }
-                }
-
-                restartApplication()
-                resetDone.value = true
             }
         }
     }
