+++ SeforimApp/src/jvmTest/kotlin/io/github/kdroidfilter/seforimapp/features/settings/general/GeneralSettingsStateTest.kt
@@ -16,7 +16,6 @@ class GeneralSettingsStateTest {
         assertFalse(state.closeTreeOnNewBook)
         assertTrue(state.persistSession)
         assertTrue(state.keepScreenAwakeOnBook)
-        assertFalse(state.resetDone)
     }
 
     @Test
@@ -26,13 +25,11 @@ class GeneralSettingsStateTest {
                 databasePath = "/custom/path/db.sqlite",
                 closeTreeOnNewBook = true,
                 persistSession = false,
-                resetDone = true,
             )
 
         assertEquals("/custom/path/db.sqlite", state.databasePath)
         assertTrue(state.closeTreeOnNewBook)
         assertFalse(state.persistSession)
-        assertTrue(state.resetDone)
     }
 
     @Test
@@ -57,7 +54,6 @@ class GeneralSettingsStateTest {
         assertEquals("/Users/you/.zayit/seforim.db", preview.databasePath)
         assertTrue(preview.closeTreeOnNewBook)
         assertTrue(preview.persistSession)
-        assertFalse(preview.resetDone)
     }
 
     @Test
